Fix sign-compare warnings in ModSecurity::processContentOffset
Cast parsed highlight offsets to size_t before comparing with content
and variable lengths.

Fixes GCC -Wsign-compare:

modsecurity.cc:271:30: warning: comparison of integer expressions of
different signedness [-Wsign-compare]

modsecurity.cc:350:30: warning: comparison of integer expressions of
different signedness [-Wsign-compare]

+++ src/modsecurity.cc
@@ -268,7 +268,7 @@ int ModSecurity::processContentOffset(const char *content, size_t len,
             size.size());
         yajl_gen_map_close(g);
 
-        if (stoi(startingAt) >= len) {
+        if (static_cast<size_t>(stoi(startingAt)) >= len) {
             *err = "Offset is out of the content limits.";
             return -1;
         }
@@ -347,7 +347,7 @@ int ModSecurity::processContentOffset(const char *content, size_t len,
             size.size());
         yajl_gen_map_close(g);
 
-        if (stoi(startingAt) >= varValue.size()) {
+        if (static_cast<size_t>(stoi(startingAt)) >= varValue.size()) {
             *err = "Offset is out of the variable limits.";
             return -1;
         }
